LPS-184766 HttpServiceRuntimeImpl: SF, split implementation of HttpServiceRuntime and other logic into 2 classes so that we don't pass an impl class everywhere.
