+++ modules/apps/static/portal-osgi-web/portal-osgi-web-http-servlet-impl/src/main/java/com/liferay/portal/osgi/web/http/servlet/internal/customizer/ContextServletTrackerCustomizer.java
@@ -14,7 +14,7 @@ package com.liferay.portal.osgi.web.http.servlet.internal.customizer;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.osgi.web.http.servlet.internal.HttpServiceRuntimeImpl;
+import com.liferay.portal.osgi.web.http.servlet.internal.HttpServiceRuntimeController;
 import com.liferay.portal.osgi.web.http.servlet.internal.context.ContextController;
 import com.liferay.portal.osgi.web.http.servlet.internal.error.HttpWhiteboardFailureException;
 import com.liferay.portal.osgi.web.http.servlet.internal.registration.ServletRegistration;
@@ -42,10 +42,10 @@ public class ContextServletTrackerCustomizer
 
 	public ContextServletTrackerCustomizer(
 		BundleContext bundleContext,
-		HttpServiceRuntimeImpl httpServiceRuntimeImpl,
+		HttpServiceRuntimeController httpServiceRuntimeController,
 		ContextController contextController) {
 
-		super(bundleContext, httpServiceRuntimeImpl);
+		super(bundleContext, httpServiceRuntimeController);
 
 		_contextController = contextController;
 	}
@@ -70,7 +70,7 @@ public class ContextServletTrackerCustomizer
 		}
 
 		if (!_contextController.matches(serviceReference) ||
-			!httpServiceRuntimeImpl.matches(serviceReference)) {
+			!httpServiceRuntimeController.matches(serviceReference)) {
 
 			return null;
 		}
@@ -123,7 +123,7 @@ public class ContextServletTrackerCustomizer
 			registration.destroy();
 		}
 
-		httpServiceRuntimeImpl.removeFailedServletDTOs(serviceReference);
+		httpServiceRuntimeController.removeFailedServletDTOs(serviceReference);
 	}
 
 	private void _recordFailedServletDTO(
@@ -149,7 +149,7 @@ public class ContextServletTrackerCustomizer
 		failedServletDTO.servletContextId = _contextController.getServiceId();
 		failedServletDTO.servletInfo = Const.BLANK;
 
-		httpServiceRuntimeImpl.recordFailedServletDTO(
+		httpServiceRuntimeController.recordFailedServletDTO(
 			serviceReference, failedServletDTO);
 	}
